feat(dashboard): promote the catalog engine to AdminDashboard; delete the inline path
The cutover. AdminDashboard is now the catalog-driven, interactive engine
(was AdminDashboardV2): renders entirely from the MDMS catalog + DashboardPack via
useCatalog + runKpiBatch + KpiTile, with the kpiId-keyed useCatalogLayout grid
(drag/resize/remove/reset). The production mount (App.js -> ./dashboard/AdminDashboard)
picks it up unchanged.

DELETED the duplicated inline data+render path:
- AdminDashboardV2.jsx (promoted into AdminDashboard.jsx)
- hooks/useDashboardData.js (inline buildBatchQueries batch — superseded by runKpiBatch)
- components/DashboardGrid.jsx (the per-widget-id renderWidget switch — superseded by KpiTile)

Single source of truth is now the backend catalog. (Follow-up cleanup: trim dead
exports in kpiQueries/kpiDisplay/layoutConfig + retire the legacy useDashboardLayout
once useCatalogLayout's geometry helpers are relocated; re-source the Add-KPI picker.)
